Rename objects to classes
