+++ jelapi/__init__.py
@@ -8,7 +8,7 @@ hoster_domain = None
 
 from .exceptions import JelasticAPIException  # noqa
 from .jelapi import JelasticAPI  # noqa
-from .objects import JelasticEnvironment  # noqa
+from .classes import JelasticEnvironment  # noqa
 
 _api_connector = None
